Update the entire directory structure
