+++ src/components/Craft/Craft.js
@@ -1,0 +1,116 @@
+import React from 'react';
+import '../../pages/Home/Hom.css';
+
+
+function Craft() {
+
+  return (
+
+    <div >
+      <section className style={{ backgroundImage: 'url("./images/e20.jpg")', backgroundSize: 'cover', backgroundPosition: 'center' }}>
+        <div class='py-5'>
+          <div className="container">
+            <div className="col">
+              <div className="craftgradient_background">
+                <div className="container text-center">
+                  <h2 style={{ color: 'white' }}>Art and Craft</h2>
+                  <p>
+                    <em style={{ color: 'white' }}>Learn creative , become creative .</em>
+                  </p>
+                </div>
+              </div>
+              <div className="row">
+                <div className="col-md-4">
+                  <div className="py-5"></div>
+                  <div className="container">
+                    <h3 style={{ color: '#880E4F' }}>1. Wall Hanging - Bamboo Sticks</h3>
+                  </div>
+                  <div className="ratio ratio-16x9">
+                    <iframe src='https://www.youtube.com/embed/LxHbPB6usmc?si=R7Zzdo3di9qKoMhQ' title="Leadership" allowFullScreen></iframe>
+                  </div>
+                </div>
+                <div className="col-md-4">
+                  <div className="py-5"></div>
+                  <div className="container">
+                    <h3 style={{ color: '#E65100' }}>2.Home Decor- Old Diya DIY </h3>
+                  </div>
+                  <div className="ratio ratio-16x9">
+                    <iframe src="https://www.youtube.com/embed/SRCptSWG-DA?si=L3xTDbS01764kHpG" title="Stress Management" allowFullScreen></iframe>
+                  </div>
+                </div>
+                <div className="col-md-4">
+                  <div className="py-5"></div>
+                  <div className="container">
+                    <h3 style={{ color: '#01579B' }}>3.Ganesh Idol - Paper Clay</h3>
+                  </div>
+                  <div className="ratio ratio-16x9">
+                    <iframe src="https://www.youtube.com/embed/BNYcxrPr2v0?si=Awej0pbNQ2gd_maC" title="Critical Thinking" allowFullScreen></iframe>
+                  </div>
+                </div>
+                <div className="col-md-4">
+                  <div className="py-5"></div>
+                  <div className="container">
+                    <h3 style={{ color: '#827717' }}>4. Pen-Pencil holder - Waste Plastic bottles </h3>
+                  </div>
+                  <div className="ratio ratio-16x9">
+                    <iframe src="https://www.youtube.com/embed/VdohdU2uOBQ?si=T591z6HHAdJ6BgyV" title="Self-Awareness" allowFullScreen></iframe>
+                  </div>
+                </div>
+                <div className="col-md-4">
+                  <div className="py-5"></div>
+                  <div className="container">
+                    <h3 style={{ color: '#E57373' }}>5. Wall paintings - Broken Bangles</h3>
+                  </div>
+                  <div className="ratio ratio-16x9">
+                    <iframe src="https://www.youtube.com/embed/0aOLpHASf08?si=QAyZDYa0gB9vGqww" title="Time Management" allowFullScreen></iframe>
+                  </div>
+                </div>
+                <div className="col-md-4">
+                  <div className="py-5"></div>
+                  <div className="container">
+                    <h3 style={{ color: '#3E2723' }}>6. Beautiful Lantern - Waste items</h3>
+                  </div>
+                  <div className="ratio ratio-16x9">
+                    <iframe src="https://www.youtube.com/embed/BJuru6ozsIQ?si=-GEwVPxQsbmlWivA" title="Time Management" allowFullScreen></iframe>
+                  </div>
+                </div>
+                <div className="col-md-4">
+                  <div className="py-5"></div>
+                  <div className="container">
+                    <h3 style={{ color: 'blue' }}>7. Garden Planters - Coconut shells</h3>
+                  </div>
+                  <div className="ratio ratio-16x9">
+                    <iframe src="https://www.youtube.com/embed/ZHZ2KgXCwRE?si=Rv0zGyS5cErmDlnH" title="Time Management" allowFullScreen></iframe>
+                  </div>
+                </div>
+                <div className="col-md-4">
+                  <div className="py-5"></div>
+                  <div className="container">
+                    <h3 style={{ color: '#006064' }}>8. Table Decor's - Waste LED Bulbs </h3>
+                  </div>
+                  <div className="ratio ratio-16x9">
+                    <iframe src="https://www.youtube.com/embed/1or0k9TvsAc?si=4v-dTHYYbnKlDHJO" title="Time Management" allowFullScreen></iframe>
+                  </div>
+                </div>
+                <div className="col-md-4">
+                  <div className="py-5"></div>
+                  <div className="container">
+                    <h3 style={{ color: '#FF1744' }}>9.Flower Vase - Plastic Spoons</h3>
+                  </div>
+                  <div className="ratio ratio-16x9">
+                    <iframe src="https://www.youtube.com/embed/a2mT6FNWTOM?si=bUtmsT4OVUfH16FG" title="Time Management" allowFullScreen></iframe>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </section>
+      <footer style={{ backgroundColor: '#333', color: '#fff', padding: '20px', textAlign: 'center' }}>
+        <p>&copy; Edu@enyanjyoti. All rights reserved.</p>
+      </footer>
+    </div>
+  );
+}
+
+export default Craft;
