+++ src/components/Education/Educ.js
